app moved to components directory
