+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
-import App from './containers/App';
+import App from './components/App';
 import store from './store';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './assets/scss/styles.scss';
